gst/sdp/gstsdpdemux.h: Change signature of pt to fix compilation on some platforms.
Original commit message from CVS:
* gst/sdp/gstsdpdemux.h:
Change signature of pt to fix compilation on some platforms.

+++ gst/sdp/gstsdpdemux.h
@@ -74,7 +74,7 @@ struct _GstSDPStream {
   GstPad       *rtcppad;
 
   /* state */
-  guint8        pt;
+  gint          pt;
   gboolean      container;
 };
